requirements.txt
# Core ML/Data Science
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0

# MLOps and Experiment Tracking  
mlflow>=2.8.0
prefect>=2.14.0

# Web Service
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
streamlit>=1.28.0
plotly>=5.17.0

# Infrastructure
boto3>=1.29.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
bandit>=1.7.0
requests>=2.31.0

# Data Processing
hyperopt>=0.2.7
evidently[ui]>=0.4.11
yfinance>=0.2.0
pyarrow>=13.0.0
